fix: fix profile layout
