+++ src/components/feed-order-details/styles.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  margin: 122px 0 122px 0;
+  margin: 122px 0 0;
   color: white;
   box-sizing: border-box;
   text-decoration: none;
