+++ src/components/feed-order-details/styles.module.scss
@@ -4,7 +4,7 @@ $circular_image_size: 64px;
   display: flex;
   flex-direction: column;
   justify-content: space-between;
-  margin: 122px 0 122px 0;
+  margin: 122px 0 0;
   color: white;
   box-sizing: border-box;
   text-decoration: none;
