+++ src/components/feed-order/styles.module.scss
@@ -1,0 +1,5 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
