+++ src/pages/profile-layout/styles.module.css
@@ -1,0 +1,11 @@
+.wrapper {
+  display: grid;
+  grid-template-columns: 320px 1fr;
+  width: 100%;
+  gap: 60px;
+  margin-top: 120px;
+}
+
+.list {
+  margin-bottom: 80px;
+}/*# sourceMappingURL=styles.module.css.map */
