+++ src/pages/profile/styles.module.css
@@ -1,0 +1,32 @@
+.container .left {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+.container .list {
+  margin-bottom: 80px;
+}
+.container .list li:not(:last-child) {
+  padding-bottom: 20px;
+}
+
+.text {
+  composes: text text_type_main-medium text_color_inactive pb-6 from global;
+  cursor: pointer;
+}
+
+.active {
+  composes: text text_type_main-medium pb-6 from global;
+  cursor: pointer;
+}
+
+.inActive {
+  composes: text text_type_main-medium text_color_inactive pb-6 from global;
+  cursor: pointer;
+}
+
+form .wrapper {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+}/*# sourceMappingURL=styles.module.css.map */
