+++ src/pages/profile/styles.module.scss
@@ -1,12 +1,4 @@
 .container {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  max-width: 860px;
-  width: 100%;
-  gap: 60px;
-  margin-top: 120px;
-  align-self: flex-start;
-
   .left {
     display: flex;
     flex-direction: column;
